requirements.txt
alembic==1.10.4
amqp==5.1.1
billiard==3.6.4.0
blinker==1.6.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.3.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
Pillow==9.5.0
prompt-toolkit==3.0.38
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.11
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.2
